requirements.txt
python-keycloak-client==0.1.3
python-keycloak==0.16.0
PyJWT==2.8.0
